+++ src/Application/Cli.php
@@ -176,6 +176,15 @@ final class Cli implements Implementation
      * @psalm-mutation-free
      */
     #[\Override]
+    public function mapRoute(callable $map): self
+    {
+        return $this;
+    }
+
+    /**
+     * @psalm-mutation-free
+     */
+    #[\Override]
     public function notFoundRequestHandler(callable $handle): self
     {
         return $this;