+++ src/Application/Http.php
@@ -40,6 +40,7 @@ final class Http implements Implementation
      * @param \Closure(OperatingSystem, Environment): Builder $container
      * @param Sequence<callable(Pipe, Container, OperatingSystem, Environment): Component<SideEffect, Response>> $routes
      * @param \Closure(RequestHandler, Container, OperatingSystem, Environment): RequestHandler $mapRequestHandler
+     * @param \Closure(Component<SideEffect, Response>, Container): Component<SideEffect, Response> $mapRoute
      * @param Maybe<callable(ServerRequest, Container, OperatingSystem, Environment): Response> $notFound
      */
     private function __construct(
@@ -48,6 +49,7 @@ final class Http implements Implementation
         private \Closure $container,
         private Sequence $routes,
         private \Closure $mapRequestHandler,
+        private \Closure $mapRoute,
         private Maybe $notFound,
     ) {
     }
@@ -66,6 +68,7 @@ final class Http implements Implementation
             static fn() => Builder::new(),
             Sequence::lazyStartingWith(),
             static fn(RequestHandler $handler) => $handler,
+            static fn(Component $component) => $component,
             $notFound,
         );
     }
@@ -83,6 +86,7 @@ final class Http implements Implementation
             $this->container,
             $this->routes,
             $this->mapRequestHandler,
+            $this->mapRoute,
             $this->notFound,
         );
     }
@@ -100,6 +104,7 @@ final class Http implements Implementation
             $this->container,
             $this->routes,
             $this->mapRequestHandler,
+            $this->mapRoute,
             $this->notFound,
         );
     }
@@ -121,6 +126,7 @@ final class Http implements Implementation
             ),
             $this->routes,
             $this->mapRequestHandler,
+            $this->mapRoute,
             $this->notFound,
         );
     }
@@ -155,6 +161,7 @@ final class Http implements Implementation
             $this->container,
             ($this->routes)($handle),
             $this->mapRequestHandler,
+            $this->mapRoute,
             $this->notFound,
         );
     }
@@ -183,6 +190,29 @@ final class Http implements Implementation
                 $os,
                 $env,
             ),
+            $this->mapRoute,
+            $this->notFound,
+        );
+    }
+
+    /**
+     * @psalm-mutation-free
+     */
+    #[\Override]
+    public function mapRoute(callable $map): self
+    {
+        $previous = $this->mapRoute;
+
+        return new self(
+            $this->os,
+            $this->env,
+            $this->container,
+            $this->routes,
+            $this->mapRequestHandler,
+            static fn($component, $get) => $map(
+                $previous($component, $get),
+                $get,
+            ),
             $this->notFound,
         );
     }
@@ -199,6 +229,7 @@ final class Http implements Implementation
             $this->container,
             $this->routes,
             $this->mapRequestHandler,
+            $this->mapRoute,
             Maybe::just($handle),
         );
     }
@@ -209,10 +240,12 @@ final class Http implements Implementation
         $container = ($this->container)($this->os, $this->env)->build();
         $os = $this->os;
         $env = $this->env;
+        $mapRoute = $this->mapRoute;
         $pipe = Pipe::new();
-        $routes = $this->routes->map(
-            static fn($handle) => $handle($pipe, $container, $os, $env),
-        );
+        $routes = $this
+            ->routes
+            ->map(static fn($handle) => $handle($pipe, $container, $os, $env))
+            ->map(static fn($component) => $mapRoute($component, $container));
         $router = new Router(
             $routes,
             $this->notFound->map(