+++ src/Application/Implementation.php
@@ -102,6 +102,15 @@ interface Implementation
     /**
      * @psalm-mutation-free
      *
+     * @param callable(Component<SideEffect, Response>, Container): Component<SideEffect, Response> $map
+     *
+     * @return self<I, O>
+     */
+    public function mapRoute(callable $map): self;
+
+    /**
+     * @psalm-mutation-free
+     *
      * @param callable(ServerRequest, Container, OperatingSystem, Environment): Response $handle
      *
      * @return self<I, O>